+++ RELEASE_NOTES.md
@@ -10,5 +10,5 @@ This is the offical release of vpinfe v1.1.12.
 None
 
 ### Notes
-None
-
+- Removed the DOF config dir clear, and instead overrite.
+- add link to online dof config tool